// apps/web/src/lib/generated/models/GovernanceFactOriginRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GovernanceFactOriginRead = {
  collected_at: string;
  model_name: string;
  model_provider: string;
  source_content_sha256: string;
  source_file_name: string;
  source_name: string;
  source_version_number: number;
};